fix output elem size, DRY shard helper, DRAM fallback test
