+++ tests/ttnn/utils_for_testing.py
@@ -879,3 +879,85 @@ def assert_div_by_zero_outputs(
     if finite_mask.any():
         # Safety net: not reached when golden is all ±inf after zero replacement.
         assert_with_ulp(golden_tensor[finite_mask], device_tensor[finite_mask], ulp_threshold=ulp_threshold)
+
+
+# ---------------------------------------------------------------------------
+# Sharded memory-config helpers (shared by universal-input TM test suites)
+# ---------------------------------------------------------------------------
+
+_DTYPE_ELEM_SIZE = {
+    ttnn.bfloat16: 2,
+    ttnn.float32: 4,
+    ttnn.uint32: 4,
+    ttnn.int32: 4,
+    ttnn.uint16: 2,
+    ttnn.uint8: 1,
+    ttnn.bfloat8_b: 1,
+}
+
+
+def _divisible_grid_1d(total_dim, max_cores, step):
+    """Return the largest n <= max_cores such that total_dim % (n * step) == 0."""
+    for n in range(max_cores, 0, -1):
+        if total_dim % (n * step) == 0:
+            return n
+    raise ValueError(f"No valid 1D grid size for total_dim={total_dim}, max_cores={max_cores}, step={step}")
+
+
+def make_sharded_memory_config(device, shape, strategy, layout, dtype=ttnn.bfloat16, buffer_type=ttnn.BufferType.L1):
+    """Create a valid sharded MemoryConfig for `shape` on `device`.
+
+    For TILE layout the shard dims must be tile-aligned (multiples of 32).
+    For ROW_MAJOR, shard_width * element_size must be a multiple of the
+    recommended L1 alignment (64 bytes today), so the shard-width step is
+    derived from the `dtype` argument.
+
+    `buffer_type` defaults to L1 (via create_sharded_memory_config). Pass
+    ttnn.BufferType.DRAM to build an equivalent DRAM-sharded config.
+    """
+    grid = device.compute_with_storage_grid_size()
+    max_x, max_y = grid.x, grid.y
+    tile_h, tile_w = 32, 32
+
+    shape_for_memcfg = list(shape)
+    if layout == ttnn.TILE_LAYOUT and len(shape) >= 2:
+        padded_h_dim = ((shape[-2] + tile_h - 1) // tile_h) * tile_h
+        padded_w_dim = ((shape[-1] + tile_w - 1) // tile_w) * tile_w
+        total_h = padded_h_dim
+        for d in shape[:-2]:
+            total_h *= d
+        total_w = padded_w_dim
+        shape_for_memcfg[-2] = padded_h_dim
+        shape_for_memcfg[-1] = padded_w_dim
+    else:
+        total_h = 1
+        for d in shape[:-1]:
+            total_h *= d
+        total_w = shape[-1]
+
+    step_h = tile_h if layout == ttnn.TILE_LAYOUT else 1
+    recommended_alignment_bytes = 64
+    element_size = _DTYPE_ELEM_SIZE.get(dtype, 2)
+    rm_step_w = max(1, recommended_alignment_bytes // element_size)
+    step_w = tile_w if layout == ttnn.TILE_LAYOUT else rm_step_w
+
+    if strategy == ttnn.ShardStrategy.HEIGHT:
+        ny = _divisible_grid_1d(total_h, max_y, step_h)
+        core_grid = ttnn.CoreGrid(y=ny, x=1)
+    elif strategy == ttnn.ShardStrategy.WIDTH:
+        nx = _divisible_grid_1d(total_w, max_x, step_w)
+        core_grid = ttnn.CoreGrid(y=1, x=nx)
+    else:  # BLOCK
+        ny = _divisible_grid_1d(total_h, max_y, step_h)
+        nx = _divisible_grid_1d(total_w, max_x, step_w)
+        core_grid = ttnn.CoreGrid(y=ny, x=nx)
+
+    l1_cfg = ttnn.create_sharded_memory_config(
+        shape=shape_for_memcfg,
+        core_grid=core_grid,
+        strategy=strategy,
+        orientation=ttnn.ShardOrientation.ROW_MAJOR,
+    )
+    if buffer_type == ttnn.BufferType.L1:
+        return l1_cfg
+    return ttnn.MemoryConfig(l1_cfg.memory_layout, buffer_type, l1_cfg.shard_spec)
